Use .gist for TTY output, .perl otherwise
With TTY output, readability is first and foremost. Because of how it's
formatted. This is not the case when $*TRACER is not a TTY; if you're
tracing to a file, gist methods like Str.gist prevent the output from
being possible to parse with a grammar, and you likely care more about
being able to recreate objects traced when this is the case anyway.