Introduce trace categories, make trace formatting easier to read/parse
Lookups/binds/assignments are not a concept exclusive to stashes. This
means that it would be very difficult to differentiate between a stash
trace and another type of trace that involves lookups/binds/assignments
when using a grammar to parse trace output, thus categories of traces
are necessary. This is too much information to include in a trace using
the old way of formatting it though.

This moves trace types (and the new categories) from the trace's header
to the trace's title. In addition, routine call headers now include the
package the routine belongs to after the routine's declarator, rather
than before it. This, I feel, makes it easier to grok trace output while
making it possible for types of traced events to not be exclusive to one
type of traced object.