Optimize and fix parameter to argument mapping in Traced::Routine
Traced::Routine.parameters-to-arguments used a gather to handle
parameter mapping. This isn't really necessary, as it can be written to
use Iterator fairly easily. Doing so makes routine tracing roughly 12.5%
faster.

Besides this, Named parameters weren't being handled properly, as
arguments corresponding to aliases of named parameters in a routine's
signature would never be mapped. This is now fixed.