Redesign the Traced/Tracer API
The v0.x API for generating traces is impossible to extend in an
ecosystem context because Tracer and Traced are codependent. This cannot
be fixed without drastic changes to these types, such that they no
longer have the same responsibilities they did before.

Unlike the old API, the new one can be summed up in a sentence: Tracers
render Traced events. Traced events should call $*TRACER.render with a
Traced instance containing any information that can be sniffed from the
event however it can. Traced no longer cares what a Tracer chooses to do
with it, so Tracer is dependent on Traced, but not vice versa.