Abstract Standard, File, and Stream tracers from Tracer::Default
Tracer::Default assumes that people will want human-readable output when
using TTYs and machine-readable output for files. The problem with this
is people may redirect output to another file or may want to save
human-readable output for later. For this reason, Tracer::File and
Tracer::Stream have been introduced.

Another problem is I/O is rather slow. People may want an in-memory
tracer! Thus the existence of Tracer::Standard, a stringy tracer with
the standard format Trait::Traced uses.