Introduce the concept of tracees
Tracer was quirky in ways that made tracers impossible to extend from an
ecosystem context. This was in part due Tracer::Standard making
transformations of traces to output unique.

This introduces a third wheel to Tracer and Traced: Tracee. This is a
generic means of transforming traces to a format tracers can output.
For now, Tracee::Bitty and Tracee::Pretty replace Tracer::Standard. With
Tracee, further simplification of the later stages of tracing can be
made with no cost in performance.