Make tests robust against proto semantic changes
Raku implementations are allowed to elide calling an onlystar proto so
long as they can be sure any signature type constraints will be met
(always the case in one with a signature of `|`). Rakudo at the time of
writing has rather "interesting" semantics here; whether it elided the
`proto` or not depended on the content of the multi dispatch cache at
the time of the call! It would always make at least one call to a proto
that was in theory possible to elide for the cache setup phase.

By contrast, the new dispatch mechanism handles multiple dispatch in a
rather different way, caching at callsites, and without any calls to an
onlystar proto with a signature it can prove may be elided. While there
was no spec against the former, the latter certainly seems the more
predictable!

This module depends on the current implementation detail of the `proto`
being called once even if it wouldn't be again. Thankfully, it's easy to
adapt it to the upcoming dispatch changes by making the `proto`s used in
tests not be onlystar ones, which is what this commit does.